opus-excel-vba-py-converter/pyproject.toml
[project]
name = "vba-to-python-converter"
version = "1.0.0"
description = "LLM-powered Excel VBA to Python code converter"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["vba", "python", "excel", "converter", "llm", "claude", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "flask>=2.3.0",
    "werkzeug>=2.3.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "oletools>=0.60.1",
    "olefile>=0.46",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    "streamlit>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
vba-converter = "app:main"

[project.urls]
Homepage = "https://github.com/yourusername/opus-excel-vba-py-converter"
Repository = "https://github.com/yourusername/opus-excel-vba-py-converter"
Documentation = "https://github.com/yourusername/opus-excel-vba-py-converter#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=html"
